Don’t try to delete non-existing files

+++ src/Content/SocialImage.php
@@ -63,8 +63,10 @@ class SocialImage
         $filePath = collect(Folder::getFilesRecursively($cachePath))
             ->firstWhere(fn ($path) => str_contains($path, $this->path()));
 
-        // Delete the cached image.
-        StatamicFile::delete($filePath);
+        if ($filePath) {
+            // Delete the cached image.
+            StatamicFile::delete($filePath);
+        }
 
         // Clean up subfolders.
         Folder::deleteEmptySubfolders($cachePath);
